Implement jmp

+++ src/cpu/opcodes/addressing_mode/indirect/mod.rs
@@ -4,5 +4,5 @@ use memory::Memory;
 pub struct Indirect;
 
 impl<M: Memory> AddressingMode<M> for Indirect {
-    type Output = u8;
+    type Output = u16;
 }
